Moved RaidenError tests to utils
